Move new method to abstract superclass
